ISDB-235: Home phone and mobile phone number implementation screen
