+++ src/main/steps/edge-case/email-address/content.ts
@@ -1,0 +1,56 @@
+import { EmailAddress } from '../../../app/case/definition';
+import { TranslationFn } from '../../../app/controller/GetController';
+import { FormContent } from '../../../app/form/Form';
+import { isEmailValid } from '../../../app/form/validation';
+import { ResourceReader } from '../../../modules/resourcereader/ResourceReader';
+
+export const form: FormContent = {
+  fields: {
+    emailAddress: {
+      type: 'text',
+      classes: 'govuk-input',
+      label: l => l.label,
+      hint: h => h.hint,
+
+      values: [{ label: l => l.emailAddress, value: EmailAddress.EMAIL_ADDRESS }],
+      validator: value => isEmailValid(value),
+    },
+  },
+  submit: {
+    text: l => l.continue,
+  },
+};
+
+export const generateContent: TranslationFn = content => {
+  const resourceLoader = new ResourceReader();
+  resourceLoader.Loader('email-address');
+  const Translations = resourceLoader.getFileContents().translations;
+  const errors = resourceLoader.getFileContents().errors;
+
+  const en = () => {
+    return {
+      ...Translations.en,
+      errors: {
+        ...errors.en,
+      },
+    };
+  };
+  const cy = () => {
+    return {
+      ...Translations.cy,
+      errors: {
+        ...errors.cy,
+      },
+    };
+  };
+
+  const languages = {
+    en,
+    cy,
+  };
+  const translations = languages[content.language]();
+  return {
+    ...translations,
+    form,
+  };
+};
