+++ src/main/steps/urls.ts
@@ -34,8 +34,7 @@ export const FIND_ADDRESS: PageLink = '/address/lookup';
 export const SELECT_ADDRESS: PageLink = '/address/select';
 export const EMAIL_ADDRESS: PageLink = '/email-address';
 export const CONTACT_NUMBER: PageLink = '/contact-number';
-export const MANUAL_ADDRESS: PageLink = '/address/manual';
-//export const CONTACT_DETAILS: PageLink = '/contact-details';         //RB
+export const MANUAL_ADDRESS: PageLink = '/address/manual';        //RB
 export const PAY_YOUR_FEE: PageLink = '/payment/pay-your-fee';
 export const UPLOAD_YOUR_DOCUMENTS: PageLink = '/upload-your-documents';
 export const CONTACT_PREFERENCES: PageLink = '/contact-preferences';
